layout/style/test/test_flexbox_layout.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=666041
-->
<head>
  <title>Test for Bug 666041</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="db_flexbox_flex.js"></script>
  <script type="text/javascript" src="property_database.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=666041">Mozilla Bug 666041</a>
<p id="display"></p>
<div id="content">
</div>
<pre id="test">
<script type="application/javascript;version=1.7">

/** Test for Bug 666041 **/
SimpleTest.waitForExplicitFinish();

function getComputedStyleWrapper(elem, prop)
{
  return window.getComputedStyle(elem, null).getPropertyValue(prop);
}

function testFlexboxTestcase(aFlexboxTestcase)
{
  let content = document.getElementById("content");

  // Create flexbox container
  let flexbox = document.createElement("div");
  flexbox.style.display = "-moz-flex";
  flexbox.style.width = ("flexboxWidth" in aFlexboxTestcase ?
                         aFlexboxTestcase.flexboxWidth : gDefaultFlexboxSize);
  flexbox.style.height = ("flexboxHeight" in aFlexboxTestcase ?
                          aFlexboxTestcase.flexboxHeight : gDefaultFlexboxSize);

  // Create & append flexbox items
  aFlexboxTestcase.items.forEach(function(aChildSpec) {
    // XXXdholbert should probably allow for other child element tagnames
    let tagName = ("tagName" in aChildSpec ?
                   aChildSpec.tagName : "div");
    let child = document.createElement(tagName);

    for (propName in aChildSpec) {
      let value = Array.isArray(aChildSpec[propName]) ?
        aChildSpec[propName][0] :
        aChildSpec[propName];

      // SANITY CHECK:
      if (Array.isArray(aChildSpec[propName])) {
        if (aChildSpec[propName].length == 3) {
          is(aChildSpec[propName][2], null, "unexpected 3rd element in array");
        } else {
          is(aChildSpec[propName].length, 2,
             "unexpected number of elements in array within child spec");
        }
      }
      if (!gCSSProperties[propName]) {
        ok(false, "Bug in test: property '" + propName +
                  "' doesn't exist in gCSSProperties.");
      } else if (value !== null) {
        var domPropName = gCSSProperties[propName].domProp;
        child.style[domPropName] = value;
      }
    }
    flexbox.appendChild(child);
  });

  content.appendChild(flexbox);
  let child = flexbox.firstChild;
  for (let i = 0; i < aFlexboxTestcase.items.length; i++) {
    if (!child) { // sanity
      ok(false, "should have created a child for each child-spec");
    }

    let childSpec = aFlexboxTestcase.items[i];
    for (propName in childSpec) {
      if (Array.isArray(childSpec[propName])) {
        let expectedVal = childSpec[propName][1];
        // Convention: any array-entry after expected value indicates 'todo'.
        var myIs = (childSpec[propName].length == 2 ? is : todo_is);
        myIs(getComputedStyleWrapper(child, propName),
             childSpec[propName][1],
             "computed value of '" + propName + "' should match expected");
      }
    }

    child = child.nextSibling;
  }
  
  content.removeChild(flexbox);
}

function main()
{
  gFlexboxTestcases.forEach(testFlexboxTestcase);
  SimpleTest.finish();
}

window.addEventListener("load", main, false);
</script>
</pre>
</body>
</html>
